# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": "Person API"
  },
  "servers": [
    {
      "url": "http://localhost:30083"
    }
  ],
  "paths": {
    "/person/list": {
      "get": {
        "summary": "Gets all people",
        "description": "",
        "responses": {
          "200": {
            "description": "Returns People",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/People"
                }
              }
            }
          }
        }
      }
    },
    "/person/samples/{count}": {
      "post": {
        "summary": "Makes N sample people and adds them to DATA (default: 5)",
        "description": "",
        "parameters": [
          {
            "name": "count",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Creates {count} people",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/People"
                }
              }
            }
          }
        }
      }
    },
    "/person/{id}": {
      "get": {
        "summary": "Gets person by ID or 404 w. Status",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns PERSON",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Person"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Deletes an existing PERSON by ID",
        "description": "Returns status message",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            }
          }
        }
      }
    },
    "/person/": {
      "post": {
        "summary": "Add a new PERSON",
        "description": "Returns Status message",
        "responses": {
          "201": {
            "description": "Created w. Status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            }
          },
          "400": {
            "description": "Bad Person w. Status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Person"
              }
            }
          }
        }
      },
      "put": {
        "summary": "Updates an existing PERSON",
        "description": "Returns status message",
        "responses": {
          "200": {
            "description": "Updated w. Status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            }
          },
          "400": {
            "description": "Bad Person w. Status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Person"
              }
            }
          }
        }
      }
    },
    "/about": {
      "get": {
        "summary": "About this project",
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Person": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "unique identity"
          },
          "firstname": {
            "type": "string",
            "example": "first name"
          },
          "lastname": {
            "type": "string",
            "example": "last name"
          },
          "cellphone": {
            "type": "string",
            "example": "cell phone"
          },
          "company": {
            "type": "string",
            "example": "company"
          },
          "email": {
            "type": "string",
            "example": "email"
          }
        },
        "required": [
          "id",
          "firstname",
          "lastname"
        ]
      },
      "People": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "array"
          },
          "items": {
            "$ref": "#/components/schemas/Person"
          }
        }
      },
      "Status": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "status message"
          },
          "data": {
            "type": "string",
            "example": "any data"
          }
        },
        "required": [
          "message"
        ]
      }
    }
  }
}